Assignment 1
Hi Instructors, I tried to push my main file a few times, but I could not submit as the .ipynb file is too large to submit (100mb is the limit, mine is ~250mb). I tried to see in a lot of places on what was causing this and I've been executing cell-by-cell to see which is been creating a lot of storage. It seems that section 4 was the largest. I am so sorry for causing confusion in this submission by sending a .zip file. This was the only way I could send in my work. I will make sure I run problems like these by you all before the next submission a few days before the due date. Sorry & thank you for your patience.